refactor: project_member Role enum
