+++ src/api/mod.rs
@@ -3,6 +3,7 @@ pub mod event;
 pub mod project;
 pub mod health_check;
 pub mod trace;
+pub mod project_member;
 
 pub use crate::api::auth::{register, login, refresh_token, get_me};
 pub use crate::api::event::{get_project_events, list_project_events, report_event, report_batch_events, set_priority, set_assignee, set_event_status};
